Add a 'scan_keys' option to replace keys() with scan().
Although 'keys' is (considerably) faster than 'scan' it can be cause
of some (considerable) long-tail latency issues.

Whereas 'scan' is (definitely) slower to gather a set of matching
keys, it does so that other clients are able to access the redis
server without some idiot blocking it with a huge 'keys' command.